Let the frontend have the graph
It was confusing that the permissions view would only show direct
permissions, instead of letting you see all permissions. Instead of
walking the graph every time we needed this data, let's just load the
graph into the frontend worker with a customizable refresh.

This is potentially adding a lot of data/load to the frontend, and
now every fe worker has to load the graph. In my testing this seems
reasonable. There are things we can do to avoid the FE loading extra
metadata if we need to.