Async notifications
This is an initial implementation of an AsyncNotification class.
Basically, you can call send_async_email and specify a time/key and then
an email will be scheduled to go out later.

This adds a thread to the API server that causes it to send emails from
the queue every minute or so and then mark them sent.

This makes all "synchronous" emails go through the async path with a
time of now, which means we get to maintain only a single path.