Make minor tweaks to the Public Key and Counter schemas to fix MySQL support
Attempting to autogenerate the SQL tables using grouper-ctl sync_db fails on MySQL systems.
This is due to the lack of length for the name column of the Counter model (this field is
a String type which translates to a SQL VARCHAR, which must have a length for MySQL). We
simply add a sufficiently large length (255) to this column in order to resolve the issue.
Additionally, the public_key column (type Text) of the PublicKey model was marked as unique.
MySQL doesn't support unique constraints on columns that do not have a specified length, so
we remove the unique constraint on the public_key column and instead move it to the fingerprint
column, which does have a maximum length, and should also be unique.

Signed-off-by: Tyler O'Meara <tyleromeara@dropbox.com>