Fix group url typos in the request update email templates
I opted to retain the use of "group" as the template variable
for the requested group name rather than the "requested" name
used in the pending_request template because I think group is
more readable. It might be a good idea to change the name of
the variable in pending_request to sync it with these
templates.