Add support for default expiration times for groups
Adds the auto_expire attribute to Group. When a user requests to
join a group (only via the front end), if they do not provide an
expiration date, their expiration is automatically set to auto_expire
days in the future. This happens for both canask and canjoin requests,
but does not apply to when a user is added to a group via a non
front end mechanism, or when an approver directly adds a member to
the group.