Add models for public key tagging
Adds the PublicKeyTag model, which represents the tags themselves.
Adds the PublicKeyTagMap model, which is used to map PublicKeyTags
to PublicKeys (since the relationship from PublicKeyTag to PublicKey
is many to many).
Adds the TagPermissionMap model, which is used to map permissions
and arguments to PublicKeyTags (since the relationship is many to
many).
The PublicKey model is updated with helper functions for easily
obtaining a PublicKey's tags, as well as the permissions that the
PublicKey has.
The AuditLog is updated to support events that occur on tags via
the addition of the on_tag relationship and on_tag_id column.