Add helper functions for interacting with PublicKeyTags
Adds the add_tag_to_public_key and remove_tag_from_public_key functions
to make adding and removing tags easier.
Adds the grant_permission_to_tag function to abstract away the process
of adding a permission to a tag.
Adds a constant for the permission required to edit the permissions
of a tag (TAG_EDIT = "grouper.tag.edit"). Permission granting access
to tags is restricted to prevent an attacker who obtains grouper web
access of a standard user from being able to add permissions to tags
that are applied to a public key that they've also compromised. This
prevents them from overriding the restrictions added by the proper
use of tags. A new permission is added rather than using the already
existing PERMISSION_GRANT permission because adding a permission to
a tag never increases a user's permissions, and therefore access does
not need to be as restricted (and requiring 50 different people to
work together to properly set up a tag would be incredibly annoying,
and thus hinder usage of tags).