Add ability to enable/disable tags.
Disabling a tag only prevents users from adding that tag to keys in the
future. It DOES NOT revoke that tag from public keys that already have
the tag, nor does it revoke permissions that were granted to that tag.
It doesn't revoke the tag from public keys because that would mean that
disabling a tag could substantially increase the permissions of one or
more public key, and that result would be very non-obvious. It doesn't
revoke permissions because if it did, every key with that tag would
become worthless.