Add support for storing password hashes for users.
Allows users to add named passwords to their accounts. The passwords
are hashed using SHA512, and salted using a unique 160-bit salt that
is randomly generated for each password. The hashes and their salts
are exposed in the users API endpoint, so high risk passwords should
not be stored in Grouper due to the lack of authentication on the
user endpoint.