Fix Permissions API for when a permission and group have the same name
There was an errant call to graph.permission_metadata.get(permission.name),
which doesn't make sense because graph permission_metadata is keyed by group
names, and only contains information relating to the mapping of all permissions
to the specified group. Additionally, this call site was expecting to receive
a dictionary, but graph.permission_metadata returns a list (or None in the
case where the provided name is not in permission_metadata). Because most
permissions do not have an identically named group, this would always
return None, and then the default of an empty dict would be returned.