Fix variable names in the password created email templates
The templates assumed that the variable called password was a password
object representing the newly created password, but it was actually
the password that was submitted by the user (as a string), so nothing
was displayed. This change makes the handler pass in just the password
name rather than the object, and changes the name in the templates
to reflect that the variable is now just the name.