Improve the performance of graph.update_from_db
Add get_all_public_key_tags to prevent us from having to use a new
database session and query for each public key (which was having a
very adverse effect on performance in graph.udpate_from_db). This also
removes the permissions key from the users public key API info, as
generating the permissions of the public keys is very slow (due to
relying on user_permissions, which is itself rather slow). It's
removal is temporary, until we are able to devise a more performant
solution.