Move template variable logic to separate function for certain views
Moved the code for retrieving and calculating the various template
variables used in UserView, GroupView, and ServiceAccountView to
separate functions on each class called get_template_vars. This
allows the template retrieval logic in ServiceAccountView to reuse
the code in UserView and GroupView, since ServiceAccountView is
essentially a combination of those two handlers.