Propagate refresh argument when we return redirects
This ensures that whatever the final respnse is, it will be based
on a fully refreshed graph. Failing to propagate the refresh could
lead to inconsistent views when using a simple loadbalancer which
directs the first request (with refresh=yes) to one server, and then
the resulting redirect request (with refresh=no) to a different
server, which has not refreshed the graph since the last change.