Force SA GroupAdds to add the group component
The UI assumes that the group component of the ServiceAccount is the
one which will be added to other groups, not the user portion. The
UI form choices reflects this by only allowing the user to see and
select the group component, but the POST handler does not see the
difference between the group and user portions, and would return
the user portion rather than the user portion. This fixes this issue
by explicitly requesting the group portion for service accounts.