Lazily evaluate public key permissions
The permissions of a public key depends on the permissions of the
user who owns the public key. Due to the way the API is constructed,
we evaluate all of the public keys when we refresh the graph.
However, user permissions are evaluated in the graph only when needed,
and are not available when the public key metadata is created. Therefore
we delay the evaluation of public key permissions until it is needed, in
order to prevent spending a significant amount of time pregenerating
information which will not be used.