Unargumented Permissions are now handled properly for tagging
Due to a typo, unargumented permissions were treated identically to
argumented permissions (arguments was being checked against None as
opposed to the empty string). However, the previous code was incorrect
in assuming that an unargumented permission should be considered a
grant wildcard (same as *). Rather, unargumented permissions should
be receive wildcards, in that any permission with the same name,
regardless of argument, grants the unargumented form.