Add test to ensure user admins can delete other user's public keys
In order to accomplish this, a new group "user-admins" is created in
the standard graph (for future use). To prevent permissions overlap
and ensure that USER_ADMIN is the reason a user can do an action,
a new user is created for this sole purpose, and added to the
user-admins group. Additionally, the USER_ADMIN permission is
added to the permissions fixture.