Add initial Secret class for secret management.
Secret management in Grouper is implemented via a plugin interface.
Grouper itself does not handle any of the secret storage, distribution,
or rotation of any secrets. It only provides an interface for users to
interact with the secrets and their metadata. In order to support as
many use cases as possible, and make integration with existing secret
management systems as easy as possible, the interface Grouper provides
for secret management is incredibly generic.

Secret management plugins implement subclasses of the Grouper provided
Secret class in order to inform Grouper about the different types of
secrets the secret management system supports. These secrets can have
arbitrary data types and names. In order to support these types, each
secret provided by the plugin must also include a subclass of SecretForm,
which is a valid form corresponding to that secret. This form class is
used to construct the form that users will use to create and edit secrets
of this type.