Add ability to view all secrets and create secrets
A new macro file is created, macros.form, for holding the form related
macros. 2 new form related macros are created, the macro 'form', which
takes an arbitrary WTForms object and renders it, and the macro
'form_modal' which is used to render forms in modals.

'form' allows us to arbitrarily render forms without having to
explicitly write out the form fields in a dedicated file as we were
doing before (see grouper.fe.templates.forms). This is essential to
support arbitrary Secret types and forms without allowing plugin
authors to write arbitrary HTML.

'form_modal' removes much of the copy pasted boiler plate we've been
using when displaying a form in a modal (a very common occurrence in
the Grouper UI). It also faciliates having multiple form modals in
a single web page, which is essential to the way the secrets_view
is implemented, because each secret type has it's own form, and thus
form modal.

A new macro 'format_secret' is added to macros.ui. This takes in
the secret object and renders is with a link, a lock symbol, and
the secrets name, similar to many of the other format macros used
such as account(), tag(), etc.

A new function 'paginate_results' is introduced in grouper.fe.util.
This function handles much the common task of taking a large list
of results and returning only a small segment of it based on
handler arguments. This allows us to remove a large amount of
boilerplate code that is common in many of the view-all-object
handlers such as secrets_view.

Finally, the typing package is added to requirements.txt, as it is
necessary for us to use the TypeVar class from that package for
type hinting 'paginate_results'