Explicitly disallow changing the type of a Secret
This was already prevented by the UI, but not strictly enforced in
the handlers. There is little to no use for this kind of feature,
and the overhead of having to try to change the type of a secret
when they can have very little in common is not worth the effort.
Additionally, if supporting changing the type of a secret is desired
in the future, it would make far more sense as a specific UI rather
than piggybacking on the edit page.