Rename the users portion of api/service_accounts
All of the collection API endpoints (groups, users, etc.) name the field
where the list of elements is stored identically to the path of the
endpoint. Prior to this change, the service accounts endpoint did not do
this, as it used the term "users", inherited from the users endpoint.
This changes it to "service_accounts" to bring it in line with the path
of the endpoint as well as the other handlers. This functionality is
expected and relied upon by Groupy.