Add Secret.get_form_args
Instead of copying the Secret object and then modifying some of the
attributes (a pretty hacky solution) in _get_secrets_form_generic, we
now have a get_form_args method for Secrets which returns the data that
the secret contains in a request args style dictionary, to be passed in
to the args attribute of WTForms to fill out the form with. This commit
also fixes a few mypy issues which were not flagged previously.