Readd the include_role_users argument to the Users API endpoint
This retains backward compatability with clients which may have been
using that argument and expecting service accounts to be returned. That
argument is considered deprecated however, and clients should migrate to
using the ServiceAccounts endpoint if they want to lookup service
accounts. This also includes some minor code review fixes.