Fix source of mysql_config during Bazel build
The build was assuming it could use the system libmysqlclient-dev
files, including the mysql_config binary, but this is no longer
true of Dropbox Bazel builds.  Instead, get mysql_config from a
Dropbox Bazel resource.

This won't work outside of Dropbox, but the Bazel configuration in
general won't, and it shouldn't impact anyone else.