Create graph and update threads after worker forking
Previously, the Graph object, the config update thread, and the graph
update thread were being created before forking out the HTTP worker
processes. This resulted in the Graph getting updated in the main
process but not any of the worker processes; so the worker processes
were serving stale data.

This commit changes it so that the Graph object and both update threads
are created after forking; so there is an independent set of these for
each worker process. This approach was taken instead of having shared
memory between the processes as it makes for a substantially simpler
implementation, and the duplicate memory impact is relatively small.

Signed-off-by: Tyler O'Meara <tomeara@quora.com>