Add smoke tests for main executables
The test suite currently doesn't catch errors in main executables
(bin/*). This adds smoke tests that call --help and implicitly check for
a clean exit. Ideally we would move more code out of these files going
forward, and test more of the functionality.

To make these tests work, we also ignore exceptions from parse_args,
which improves the UX when calling --help or when there is an argument
error.