Rename internal service account APIs to role user
Role user was the old name that was still used in places internally.
Rename all internal APIs to use `RoleName` or `role_name`.  This frees
up the `ServiceAccount` and `service_account` namespace for the new
implementation.

Do not change the API endpoints and the flags they use to specify
whether to include service accounts or not, since those will continue
working as they currently do with both old role users and new service
accounts.