Prevent removing the last owner of a group
This defines a "permanent owner" of a group as a member with a role of
owner or np-owner, and no expiration date. Any change to a group that
would remove the last permanent owner raises an exception. This is
enforced in the new GroupOwnershipPolicy plugin.

This should prevent groups from becoming ownerless in the future, but
does not affect groups which are already scheduled to become ownerless.