Suppress exceptions in the background thread
Currently, non-database related exceptions in the background thread are
reraised, terminating the background thread. This is nonideal as
temporal errors can result in the background thread ceasing to run,
requiring a server restart. This change adds logging for all unexpected
exceptions in the background thread and then suppresses the exception,
continuing to attempt to perform the background tasks.