Uncaught exceptions in the background thread exit the process
Currently, non-database related exceptions in the background thread are
reraised, terminating the background thread. This is nonideal as
temporal errors can result in the background thread ceasing to run,
requiring a server restart. This change adds logging for all unexpected
exceptions in the background thread and then raises a KeyboardInterrupt
on the main thread, causing the process to exit. This prevents the main
thread from serving requests that could be based on outdated or corrupt
information, as well as automating part of the oncall response.

Signed-off-by: Tyler O'Meara <tomeara@quora.com>