Avoid parens in test assertions
It's not safe to use parentheses to continue assertions on multiple
lines.  `assert` is a built-in, not a method call, so this passes
a tuple to `assert`, which is always true.  Remove the parentheses
and use backslash for line continuation.

Caught by py.test warnings in the Travis CI run.