Deduplicate form_field macros
There were two copies of the `form_field` macro outside of
`macros/ui.html`, and a third that had an additional parameter.
Merge all of these and get rid of duplicate code.

Remove a rows parameter for the name of a user password, which
both made no sense and wasn't being honored.