Move all code into libraries
Best practice for testable code is to put as little as possible
in the command wrapper, since one has to go to special effort to
load and run that code.  Following the pattern for grouper-ctl,
move the main() routines of grouper-api and grouper-fe into the
library.

No tests currently take advantage of this (the problem I thought
I had turned out to be something else), but it lays the groundwork
for later test work.