Fix permission argument checking
Permission argument checking was a bit of a mess.  We validated
arguments when granting permissions directly, but not when creating
a request for a permission.  The backup regex checking the argument
inside the API only checked that the argument started with a valid
string, not consisted solely of a valid string.

Add validators when requesting a permission, anchor the regex in
the assertions in the API, and add tests to ensure we're properly
validating permission arguments.

Allow spaces in permission arguments, since we were already using
them (via requested permissions).