Add missing frontend length validators
The permission argument is currently restricted to 64 characters,
and the user token name is restricted to 16 characters.  Add those
restrictions directly to the form validation, since MySQL will
otherwise accept invalid data and silently truncate it.

Note that the permission *request* table limits arguments to 64
characters, but the permission *map* table allows 128 characters,
so the limits are inconsistent between grant and request.  This will
be fixed with a later commit but will require a database migration.

Also comment all of the unconstrained choice fields that I manually
verified did provide choices at runtime.