Do not expose whether a permission is an alias in API responses
This is a bad, but hopefully temporary solution to a problem that should
not exist in the first place. Old versions of Groupy break when mapped
permissions include unexpected fields, and declaring exactly what's
included in API responses is a larger (and riskier) change. This commit
can be reverted when we no longer need to support old Groupy versions.