Don't use US/Pacific-New in tests
This is not what it sounds like.  From the tzdata source comments:

> On 1989-04-05, the U. S. House of Representatives passed (238-154) a bill
> establishing "Pacific Presidential Election Time"; it was not acted on
> by the Senate or signed into law by the President.

It's an obsolete time zone rule that was never enacted into law, so
it's not included in some environments.  Specifically, it's not available
in the Dropbox Bazel environment, so using it for the test suite is
causing test failures.