Refactor permission disabling, add grouper-ctl command
Add a permission disable command to grouper-ctl with the same
functionality as the disable permission button in the UI.

Refactor the code to introduce UI objects, service objects, and
a use case object, with interfaces between them implemented with
abstract base classes.  Plumb this particular use case from the
UI down to nearly the storage layer, except for permission
checking and a storage service with graph integration (which will
be addressed in an upcoming change).

Add a group with permissions admin to the standard graph fixture
and update the permission tests accordingly.