Add pytest-xdist to dev requirements
This allows running tests in parallel with -n.  Time to run all
tests and itests decreases to 28s from 1m 48s on my laptop.

Two audit tests relied on the standard graph but didn't explicitly
depend on it, so failed intermittantly when reordered by parallel
testing.  Fix that by adding an explicit dependency.