Code review fixes for disable permission use case
Stop trying to maintain an exit status for grouper-ctl, since it's
not currently honored anyway.  We'll have to fix this later.
Comment the janky way that we're initializing the session, and use
a better variable name for initializing subcommands.

Pass the actual Permission into the audit log method.  (Later we
will replace this with value objects from the repository layer.)