Clean up setup.py, remove paths for flake8 and pytest
setup.py wasn't flake8-clean, which was the only thing preventing
flake8 from being run without a path.  Fix that, and drop the paths
in the documentation and CI script, and also drop the similarly
unnecessary py.test paths.

Do the minimum work required to make setup.py work, namely adding
setuptools to the dependencies and telling it to use pytest to run
tests (and what the test dependencies are).  Change the syntax used
to load the version file to be compatible with Python 3 and therefore
flake8 run under Python 3.