Also run tests with MySQL, fix API session handling
We use MySQL in production, and SQLite (which is used by default
for the tests) does not do schema checking and thus won't uncover
some issues.  Add a mechanism to configure a database URL via an
environment variable, and run tests twice, once with MySQL and
once with SQLite.

Fix database session leaks in the API server uncovered by this
test.