Add factories and a repository layer
Extend the permission disable refactoring down to the storage
layer by adding repositories to isolate the rest of the code from
SQLAlchemy.  Introduce a data transfer object holding permission
details.  Add factories at the usecase, service, and repository
layer and use them to assemble objects in the UIs and tests.