Refactor listing permissions
Add a new ListPermissions use case, backed by the graph, and replace
the frontend handler code with code that calls it.

This also refactors the PermissionRepository to have two
implementations, one based on the graph and one based on SQL.  The
grouper-ctl permission command now instantiates the graph so it is
(hopefully temporarily) rather slow.

Refactor interfaces that ask about the permissions held by a user
into a separate UserService instead of the PermissionService.