Refactor API list permissions handler
Modify the ListPermissions use case to also support the API.  The
API has no current user and doesn't care about pagination, so this
involves adding a new method to the use case that doesn't take the
user, pagination, and restrict to audited permissions arguments to
hide more complexity from the caller (but uses the same API to
return results), making the limit field in pagination optional,
and adding a sort option of no sorting.

On the repository side, make sorting and slicing optional and fix
the data transfer objects returned from the repository (which were
previously accidentally type-compatible with the Permission object).