Add tests for list permissions refactor
Add a use case test and an integration test for the front end,
since there was no previous integration test of the permissions
list page.

Change the default dev timezone to UTC, since otherwise testing
date display is irritatingly complicated.  (We should later add
a test of the time zone configuration and its impact.)

Add the permission description to the data transfer object, which
apparently otherwise wasn't tested.

Add support for forcing a refresh to the permission list page,
since it makes testing easier.  Also add more class attributes to
the HTML to make Selenium testing easier.

Add support for testing the heading and subheading of a Grouper
page to the itest page framework.

Rename the existing PermissionRow to PermissionGrantRow, since
that's what it really represents, freeing PermissionRow for a row
of the permission list page.