Refactor test setup for usecase tests
Provide a new SetupTest class that handles creating the database
session and provides utility functions for setting up test data.
This is the replacement for fixtures that set up a complex data
model that most tests don't need and that somewhat obscure what
the test is doing.

This is the first iteration, which messes with the database
directly via SQLAlchemy objects.  A later revision will use the
same service and repository path that non-test code uses.

This is implemented as a fixture rather than an explicit call
because, when testing with a persistant database, we have to close
the session or the next test will hang while trying to drop the
schema to ensure fresh test data due to the leaked previous session.
Having to remember to close the session is irritating, as is having
to write every test inside a with clause.  A pytest fixture is
simplest for the test writer.

To avoid endless flake8 warnings, provide the fixture via
conftest.py so that pytest will inject it into every test without
an explicit import.  (This is somewhat spooky action at a distance,
but it saves a lot of repetitive typing.)