Refactor test setup for permission list integration test
Use the same pytest fixture setup as the tests directory in the
itests directory, and move the fixture for the browser into
conftest.py.  Use the new SetupTest support when running
permission list integration tests.

Provide a context manager to spawn the frontend server, which
allows the user to pass to the frontend as authentication to be
specified by the test, at the cost of having to write the test
inside a context manager.