Refactor test setup for permissions API integration test
Similar to the frontend test, add an api_server() context manager
to start the server and convert the permissions integration test
to use the new SetupTest mechanism rather than building the entire
graph.

The API server doesn't take a user parameter, so could be provided
as a fixture, but continue to use the context manager pattern since
it allows deterministic ordering of the test data setup before
spawning the server (and also allows stopping and restarting the
server in the same test, although this isn't used yet).