Convert transactions to a context manager
Change the API of TransactionService to create a Transaction object
that implements the context manager API, and update callers to use
it instead of start_transaction and commit pairs.  Add a Transaction
interface and a SQLTransaction implementation.  Retain the commit
method for tests, which may not want to bother with the context
manager, but hide it from use cases.