Refactor grouper-ctl initialization
Move argument setup to a static method so that it can be called
without a session, and change the way that the right class is
determined to look at args.command if args.func is not set.
Automatically initialize all discovered subclasses, so all that's
required to add a new command to ctl/main.py is to import the
relevant class.

Move the factory setup into the CtlCommand base class.

Add sys.exit calls in the failure handlers in PermissionCommand
so that the exit status is correct.