Move checkpoint updates to the transaction service
Rather than independently updating the checkpoint counter in every
change, make the transaction service update it on every transaction
commit.  Add a checkpoint service to manage the checkpoint counter.

Use SELECT FOR UPDATE to lock the counter when updating it so that
we won't potentially race with other transactions.

Add a detailed comment about why updating the last_modified column
has been disabled.