Refactor grouper-ctl command creation to use a factory
Since the Session and GroupGraph have to be optional for individual
commands, we need to keep the logic for which commands need them and
which don't somewhere.  A factory pattern collects that logic and
cleans up some other oddities, such as how to construct argument
parsers for each command.  The factory has sole knowledge of the
mapping between a command name and a CtlCommand implementation
class.