Refactor TransactionService to use a repository
Move the bits that call into SQLAlchemy into a TransactionRepository,
eliminating the last dependency of a ServiceFactory on a Session.
The TransactionService is now only responsible for ensuring a
checkpoint update, so it uses a simpler @contextmanager.  Adjust the
type of the expected interface in the use case to just any context
manager.

Add type annotations to every place a ServiceFactory was being
created so that mypy would catch the argument changes, and update
callers.  This includes removing now-duplicate work from the
PermissionDisable handler and reusing the UseCaseFactory stored in
the handler object.

Add a type signature to the post() method in the PermissionDisable
handler, which requires changing the method signature because mypy
isn't very happy with the flexibility of Tornado's handler methods.