Refactor use case factory creation
Stop requiring a session and graph be injected into RepositoryFactory
and let it create them lazily when needed, since they're both
artifacts of the repository layer.  This requires passing Settings
to the RepositoryFactory (to get the database URL information).

Rather than repeating the same three lines to create a UseCaseFactory,
factor that out into a create_usecase_factory() function in a new
grouper.initialization module, with optional injection of a Session
and GroupGraph (which is somewhat widely used right now, until the API
and FE servers are further refactored).

Refactor Tornado application creation to avoid overriding the
constructor and creating new object attributes, and instead injecting
the global application settings as additional arguments to every
handler's initialize method.  Use this to simplify application creation
and plumb the new usecase_factory through to the handlers.  Stop
plumbing the SentryProxy through, since we aren't using it anywhere.