Remove the TransactionService.commit() hack
Expose a context manager from SetupTest for transactions and use it
for all test setup as well, so that tests follow the same general
coding style as the rest of the code.  Delete the commit() method on
TransactionService and TransactionRepository so that they aren't
lingering near production code as an attractive nuisance.