Some minor cleanup of disable permission
While discussing a new use case, we decided that the "because"
in UI methods for failure just makes the method name longer without
adding clarity.  Without because, it still reads as English if you
see the underscore after "failed" as a colon.

We also decided to do all checks in the use case, rather than catching
exceptions thrown by the repository.  There is some possibility of
failure later with an exception because of a race, but the cost in
complexity of catching those race condition exceptions everywhere is
high and the value unclear, particularly for objects where deletion is
not currently supported.  Rewrite the disable permission use case to
see if the permission exists first before trying to disable it, and
then don't catch exceptions in the event of a race with deleting the
permission (particularly since deleting a permission entirely is not
currently a supported operation).