Add direct SQL user permission code
Add a pure SQL implementation of the permission grant repository
that retrieves all permissions for a user directly from the
database (takes about 150ms with current Dropbox production
configurations).

Separate the repository factory into a Graph and a SQL variation,
use the Graph variation for the frontend and API server, and use
the SQL variation for grouper-ctl.  The SQL variation will never
instantiate a graph or use any of the graph functions, thus avoiding
the graph construction overhead.