Add support for converting a user to a service account
Add a use case and a grouper-ctl command that converts a user to
a service account by first disabling it and then wrapping it in a
service account and enabling the service account.

Move the user subcommand of grouper-ctl to the new framework so
that it can run use cases, although the existing subcommands are
still implemented the old way.

Move the exceptions for unknown database entities to the
grouper.entities.* modules so that they can be shared across
repositories, services, and use cases (the latter two may need to
catch the exceptions in some situations).