Fix permission pagination limit on last page
The frontend UI was using the number of returned items as the
requested limit on the number of items, which is wrong for the
last page of a paginated list.  Return the requested limit as part
of the paginated list data structure so that the actual limit isn't
lost when there are fewer than that number of remaining items.