Rewrite grouper-ctl user_proxy
mrproxy doesn't support Python 3.  It is small enough, the changes
for porting are significant enough, and we were already overriding
enough of its logic that it made sense to replace it entirely.

Rewrite using http.server and urllib.request directly, in a way
that is compatible with both Python 2 and Python 3.  Covert it to
the new grouper-ctl class structure.