Simplify by removing method type management
mrproxy supported overriding the method and configuring the
created Request with a custom method, which required a bit of
porting between Python 2 and Python 3.  However, we weren't using
that support in Grouper (which only uses GET and POST), and both
Python 2 and Python 3 Request should autodetect that case.

Drop the complexity.  Hopefully by the time we want to support
PUT or DELETE, we can be pure Python 3.