Update documentation, tweak requirements files
Document that Python 3 is now supported.  Recommend Python 3 as
the default environment.  Update the Travis-CI integration to run
the full test suite on both Python 2 and Python 3.

Reflect the change in defaults by renaming requirements-dev.txt
to requirements-dev2.txt and requirements-dev3.txt to
requirements-dev.txt.  Install coverage under Python 3, and bump
the Groupy version to the one that works without future (which
causes all sorts of problems).

Update setup.py to reflect current maintainership and to add the
Trove classifiers for Python 2.7 and Python 3 support.